requirements.txt
PyYAML==3.13
Flask==1.0.2
werkzeug==0.14.1